Move relative-select helpers to utils.pd
relative_offset and relative_ranges_to_absolute operate only on a
dataframe's envelope columns and are used by generic DataFrameSpool
selection, but lived in io.index.query, forcing core.spool to import
the index query builder for backend-independent relative selection.
Move them beside adjust_segments/filter_df in utils.pd; the catalog and
spool both import them from there. Restores a clean dependency
direction (core -> generic pd utilities, not core -> io backend).

The catalog-pushdown test now realizes via get_contents to inspect the
composed query, since len() counts in SQL and no longer fetches rows.
